Handle query parameters
This fixes it so that query parameters are properly handled and because
ViewModel properties.

Closes #103
