+++ package.json
@@ -32,6 +32,7 @@
     "can-component": "^4.0.0-pre.0",
     "can-define": "^2.0.0-pre.19",
     "can-route": "^4.0.0-pre.0",
+    "can-route-pushstate": "^4.0.0-pre.0",
     "funcunit": "^3.1.0",
     "lodash.template": "^4.4.0",
     "steal": "^1.0.3",
